FEARTUE ✨ : Moving Group Routes to the private area (EXCEPT GROUP INDEX)
The group index in the Public Area shows an empty list if the user is not logged in.  With a CTA Button to Create or Join a Group.

If they click on the Buttons, they will be redirected to the Log In Page.

If they are logged in, they see the list of the groups they are part of.