REFACTOR 🔨 : Identified and solved the N+1 problem
In the index view I count the number of users in each group. With the previous query, the view was making a query per group to fetch all the users inside the group and then count them.

I changed the query in the controller to load the users. Now we have one single query being executed.