REFACTOR 🔨 : GroupCard Component for groups index
This component changes depending on the route intead of checking if the user is an admin or not. If they have access to the admin route, it means that they are an admin.

Managed to make it reusable by using php inside the component with request()->routeIs() and with dynamic routes.